Fix master quality gate issues
