+++ packages/grpc/src/analyze-project-server.ts
@@ -17,383 +17,48 @@
 
 import * as grpc from '@grpc/grpc-js';
 import type { Worker } from 'node:worker_threads';
-import { sonarjs as analyzeProjectProto } from './proto/analyze-project.js';
 import { debug, info } from '../../shared/src/helpers/logging.js';
-import { handleAnalyzeProjectRequest, type WorkerData } from './analyze-project-handle-request.js';
 import type {
   AnalyzeProjectWorkerInMessage,
   AnalyzeProjectWorkerOutMessage,
 } from './analyze-project-worker/messages.js';
 import {
   logMemoryConfiguration,
-  logMemoryError,
   registerGarbageCollectionObserver,
 } from './analyze-project-memory.js';
-import type {
-  AnalyzeProjectIncrementalEvent,
-  AnalyzeProjectResponse,
-  AnalyzeProjectRuntimeRequest,
-  RequestResult,
-} from './analyze-project-request.js';
 import {
   toAnalyzeProjectStreamResponse,
   toAnalyzeProjectUnaryResponse,
 } from './analyze-project-convert.js';
-
-const ANALYZE_PROJECT_SERVICE_NAME = 'sonarjs.analyzeproject.v1.AnalyzeProjectService';
-const WORKER_RESPONSE_TIMEOUT_MS = 15_000;
-const GRPC_SERVER_OPTIONS: grpc.ServerOptions = {
-  'grpc.max_receive_message_length': -1,
-  'grpc.max_send_message_length': -1,
-};
-
-type AnalyzeProjectRequest = analyzeProjectProto.analyzeproject.v1.IAnalyzeProjectRequest;
-type AnalyzeProjectStreamResponse =
-  analyzeProjectProto.analyzeproject.v1.IAnalyzeProjectStreamResponse;
-type AnalyzeProjectUnaryResponse =
-  analyzeProjectProto.analyzeproject.v1.IAnalyzeProjectUnaryResponse;
-type CancelAnalysisRequest = analyzeProjectProto.analyzeproject.v1.ICancelAnalysisRequest;
-type CancelAnalysisResponse = analyzeProjectProto.analyzeproject.v1.ICancelAnalysisResponse;
-type LeaseRequest = analyzeProjectProto.analyzeproject.v1.ILeaseRequest;
-type LeaseResponse = analyzeProjectProto.analyzeproject.v1.ILeaseResponse;
+import {
+  createAnalyzeProjectServiceDefinition,
+  failStreamingCall,
+  GRPC_SERVER_OPTIONS,
+  toGrpcError,
+  toGrpcErrorFromFailure,
+  type AnalyzeProjectRequest,
+  type AnalyzeProjectStreamResponse,
+  type AnalyzeProjectUnaryResponse,
+  type CancelAnalysisRequest,
+  type CancelAnalysisResponse,
+  type LeaseRequest,
+  type LeaseResponse,
+} from './analyze-project-server-grpc.js';
+import {
+  attachWorkerLifecycleHandlers,
+  createHandleRequestInCurrentThread,
+  createLifecycle,
+  createServerState,
+  getNextWorkerRequestId,
+  waitForWorkerCompletion,
+  type AnalyzeProjectImplementationDependencies,
+} from './analyze-project-server-lifecycle.js';
 type UnaryCompleteMessage = Extract<AnalyzeProjectWorkerOutMessage, { type: 'unary-complete' }>;
 
 type AnalyzeProjectServerResult = {
   server: grpc.Server;
   serverClosed: Promise<void>;
 };
-
-type HandleRequestInCurrentThread = (
-  request: AnalyzeProjectRuntimeRequest,
-  incrementalResultsChannel?: (result: AnalyzeProjectIncrementalEvent) => void,
-) => Promise<RequestResult<AnalyzeProjectResponse | void>>;
-
-type AnalyzeProjectServerState = {
-  analysisInProgress: boolean;
-  leaseCall: grpc.ServerDuplexStream<LeaseRequest, LeaseResponse> | null;
-  nextWorkerRequestId: number;
-  shuttingDown: boolean;
-  startupShutdownTimeout: NodeJS.Timeout | null;
-};
-
-type AnalyzeProjectServerLifecycle = {
-  scheduleStartupShutdownTimeout: () => void;
-  clearStartupShutdownTimeout: () => void;
-  requestCancel: () => Promise<boolean>;
-  shutdown: (reason: string) => Promise<void>;
-};
-
-type AnalyzeProjectServerDependencies = {
-  handleRequestInCurrentThread: HandleRequestInCurrentThread;
-  newWorkerRequestId: () => string;
-  resolveClosed: () => void;
-  server: grpc.Server;
-  state: AnalyzeProjectServerState;
-  timeout: number;
-  unregisterGarbageCollectionObserver: () => void;
-  worker?: Worker;
-};
-
-type AnalyzeProjectImplementationDependencies = {
-  handleRequestInCurrentThread: HandleRequestInCurrentThread;
-  lifecycle: AnalyzeProjectServerLifecycle;
-  newWorkerRequestId: () => string;
-  state: AnalyzeProjectServerState;
-  worker?: Worker;
-};
-
-function createAnalyzeProjectServiceDefinition(): grpc.ServiceDefinition {
-  return {
-    AnalyzeProject: {
-      path: `/${ANALYZE_PROJECT_SERVICE_NAME}/AnalyzeProject`,
-      requestStream: false,
-      responseStream: true,
-      requestSerialize: (value: AnalyzeProjectRequest) =>
-        Buffer.from(
-          analyzeProjectProto.analyzeproject.v1.AnalyzeProjectRequest.encode(value).finish(),
-        ),
-      requestDeserialize: (buffer: Buffer) =>
-        analyzeProjectProto.analyzeproject.v1.AnalyzeProjectRequest.decode(buffer),
-      responseSerialize: (value: AnalyzeProjectStreamResponse) =>
-        Buffer.from(
-          analyzeProjectProto.analyzeproject.v1.AnalyzeProjectStreamResponse.encode(value).finish(),
-        ),
-      responseDeserialize: (buffer: Buffer) =>
-        analyzeProjectProto.analyzeproject.v1.AnalyzeProjectStreamResponse.decode(buffer),
-    },
-    AnalyzeProjectUnary: {
-      path: `/${ANALYZE_PROJECT_SERVICE_NAME}/AnalyzeProjectUnary`,
-      requestStream: false,
-      responseStream: false,
-      requestSerialize: (value: AnalyzeProjectRequest) =>
-        Buffer.from(
-          analyzeProjectProto.analyzeproject.v1.AnalyzeProjectRequest.encode(value).finish(),
-        ),
-      requestDeserialize: (buffer: Buffer) =>
-        analyzeProjectProto.analyzeproject.v1.AnalyzeProjectRequest.decode(buffer),
-      responseSerialize: (value: AnalyzeProjectUnaryResponse) =>
-        Buffer.from(
-          analyzeProjectProto.analyzeproject.v1.AnalyzeProjectUnaryResponse.encode(value).finish(),
-        ),
-      responseDeserialize: (buffer: Buffer) =>
-        analyzeProjectProto.analyzeproject.v1.AnalyzeProjectUnaryResponse.decode(buffer),
-    },
-    CancelAnalysis: {
-      path: `/${ANALYZE_PROJECT_SERVICE_NAME}/CancelAnalysis`,
-      requestStream: false,
-      responseStream: false,
-      requestSerialize: (value: CancelAnalysisRequest) =>
-        Buffer.from(
-          analyzeProjectProto.analyzeproject.v1.CancelAnalysisRequest.encode(value).finish(),
-        ),
-      requestDeserialize: (buffer: Buffer) =>
-        analyzeProjectProto.analyzeproject.v1.CancelAnalysisRequest.decode(buffer),
-      responseSerialize: (value: CancelAnalysisResponse) =>
-        Buffer.from(
-          analyzeProjectProto.analyzeproject.v1.CancelAnalysisResponse.encode(value).finish(),
-        ),
-      responseDeserialize: (buffer: Buffer) =>
-        analyzeProjectProto.analyzeproject.v1.CancelAnalysisResponse.decode(buffer),
-    },
-    Lease: {
-      path: `/${ANALYZE_PROJECT_SERVICE_NAME}/Lease`,
-      requestStream: true,
-      responseStream: true,
-      requestSerialize: (value: LeaseRequest) =>
-        Buffer.from(analyzeProjectProto.analyzeproject.v1.LeaseRequest.encode(value).finish()),
-      requestDeserialize: (buffer: Buffer) =>
-        analyzeProjectProto.analyzeproject.v1.LeaseRequest.decode(buffer),
-      responseSerialize: (value: LeaseResponse) =>
-        Buffer.from(analyzeProjectProto.analyzeproject.v1.LeaseResponse.encode(value).finish()),
-      responseDeserialize: (buffer: Buffer) =>
-        analyzeProjectProto.analyzeproject.v1.LeaseResponse.decode(buffer),
-    },
-  };
-}
-
-function toGrpcError(message: string, code = grpc.status.INTERNAL): grpc.ServiceError {
-  const error = new Error(message) as grpc.ServiceError;
-  error.name = 'AnalyzeProjectServiceError';
-  error.code = code;
-  error.details = message;
-  error.metadata = new grpc.Metadata();
-  return error;
-}
-
-function failStreamingCall(
-  call:
-    | grpc.ServerWritableStream<AnalyzeProjectRequest, AnalyzeProjectStreamResponse>
-    | grpc.ServerDuplexStream<LeaseRequest, LeaseResponse>,
-  error: grpc.ServiceError,
-) {
-  call.emit('error', error);
-}
-
-function toGrpcErrorFromFailure(result: Extract<RequestResult, { type: 'failure' }>) {
-  return toGrpcError(
-    result.error.message,
-    result.reason === 'invalid_request' ? grpc.status.INVALID_ARGUMENT : grpc.status.INTERNAL,
-  );
-}
-
-async function waitForWorkerCompletion(
-  worker: Worker,
-  expectedType: AnalyzeProjectWorkerOutMessage['type'],
-  expectedRequestId: string,
-  trigger: () => void,
-  timeoutMs?: number,
-): Promise<AnalyzeProjectWorkerOutMessage> {
-  return new Promise((resolve, reject) => {
-    const timeout =
-      timeoutMs === undefined
-        ? undefined
-        : setTimeout(() => {
-            cleanup();
-            reject(new Error(`Timed out waiting for worker message '${expectedType}'`));
-          }, timeoutMs);
-
-    const onMessage = (message: AnalyzeProjectWorkerOutMessage) => {
-      if (message.type !== expectedType || message.requestId !== expectedRequestId) {
-        return;
-      }
-      cleanup();
-      resolve(message);
-    };
-
-    const onError = (error: Error) => {
-      cleanup();
-      reject(error);
-    };
-
-    const onExit = (code: number) => {
-      cleanup();
-      reject(new Error(`Worker exited with code ${code}`));
-    };
-
-    const cleanup = () => {
-      if (timeout) {
-        clearTimeout(timeout);
-      }
-      worker.off('message', onMessage);
-      worker.off('error', onError);
-      worker.off('exit', onExit);
-    };
-
-    worker.on('message', onMessage);
-    worker.on('error', onError);
-    worker.on('exit', onExit);
-    trigger();
-  });
-}
-
-function createServerState(): AnalyzeProjectServerState {
-  return {
-    analysisInProgress: false,
-    leaseCall: null,
-    nextWorkerRequestId: 0,
-    shuttingDown: false,
-    startupShutdownTimeout: null,
-  };
-}
-
-function createHandleRequestInCurrentThread(workerData: WorkerData): HandleRequestInCurrentThread {
-  return (request, incrementalResultsChannel) =>
-    handleAnalyzeProjectRequest(request, workerData, incrementalResultsChannel);
-}
-
-function getNextWorkerRequestId(state: AnalyzeProjectServerState): string {
-  state.nextWorkerRequestId += 1;
-  return String(state.nextWorkerRequestId);
-}
-
-function clearStartupShutdownTimeout(state: AnalyzeProjectServerState) {
-  if (state.startupShutdownTimeout) {
-    clearTimeout(state.startupShutdownTimeout);
-    state.startupShutdownTimeout = null;
-  }
-}
-
-async function cancelAnalysisOnShutdown(
-  handleRequestInCurrentThread: HandleRequestInCurrentThread,
-  newWorkerRequestId: () => string,
-  state: AnalyzeProjectServerState,
-  worker?: Worker,
-) {
-  if (!state.analysisInProgress) {
-    return;
-  }
-
-  if (worker) {
-    try {
-      worker.postMessage({
-        type: 'cancel',
-        requestId: newWorkerRequestId(),
-      } satisfies AnalyzeProjectWorkerInMessage);
-    } catch (e) {
-      debug(`Failed to post cancel to worker: ${e}`);
-    }
-    return;
-  }
-
-  try {
-    await handleRequestInCurrentThread({ type: 'on-cancel-analysis' });
-  } catch (e) {
-    debug(`Failed to cancel in-process analysis: ${e}`);
-  }
-}
-
-async function closeWorker(worker?: Worker) {
-  if (!worker) {
-    return;
-  }
-
-  try {
-    worker.postMessage({ type: 'close' } satisfies AnalyzeProjectWorkerInMessage);
-  } catch (e) {
-    debug(`Failed to post close to worker: ${e}`);
-  }
-
-  try {
-    await worker.terminate();
-  } catch (e) {
-    debug(`Failed to terminate worker: ${e}`);
-  }
-}
-
-function createLifecycle({
-  handleRequestInCurrentThread,
-  newWorkerRequestId,
-  resolveClosed,
-  server,
-  state,
-  timeout,
-  unregisterGarbageCollectionObserver,
-  worker,
-}: AnalyzeProjectServerDependencies): AnalyzeProjectServerLifecycle {
-  const requestCancel = async () => {
-    if (!worker) {
-      const result = await handleRequestInCurrentThread({ type: 'on-cancel-analysis' });
-      return result.type === 'success';
-    }
-
-    const requestId = newWorkerRequestId();
-    const completion = await waitForWorkerCompletion(
-      worker,
-      'cancel-complete',
-      requestId,
-      () =>
-        worker.postMessage({ type: 'cancel', requestId } satisfies AnalyzeProjectWorkerInMessage),
-      WORKER_RESPONSE_TIMEOUT_MS,
-    );
-    return completion.type === 'cancel-complete' && completion.result.type === 'success';
-  };
-
-  const closeLeaseCall = () => {
-    if (!state.leaseCall) {
-      return;
-    }
-    const currentLeaseCall = state.leaseCall;
-    state.leaseCall = null;
-    try {
-      currentLeaseCall.end();
-    } catch (e) {
-      debug(`Failed to end lease stream: ${e}`);
-    }
-  };
-
-  const shutdown = async (reason: string) => {
-    if (state.shuttingDown) {
-      return;
-    }
-    state.shuttingDown = true;
-    debug(`Shutting down gRPC analyze-project server: ${reason}`);
-    clearStartupShutdownTimeout(state);
-    closeLeaseCall();
-    unregisterGarbageCollectionObserver();
-    await cancelAnalysisOnShutdown(handleRequestInCurrentThread, newWorkerRequestId, state, worker);
-    await closeWorker(worker);
-
-    server.forceShutdown();
-    resolveClosed();
-  };
-
-  const scheduleStartupShutdownTimeout = () => {
-    if (timeout <= 0) {
-      return;
-    }
-    clearStartupShutdownTimeout(state);
-    state.startupShutdownTimeout = setTimeout(() => {
-      void shutdown('lease acquisition timeout');
-    }, timeout);
-  };
-
-  return {
-    scheduleStartupShutdownTimeout,
-    clearStartupShutdownTimeout: () => clearStartupShutdownTimeout(state),
-    requestCancel,
-    shutdown,
-  };
-}
 
 function createAnalyzeProjectStreamHandler({
   handleRequestInCurrentThread,
@@ -674,27 +339,6 @@ function createAnalyzeProjectImplementation(
   };
 }
 
-function attachWorkerLifecycleHandlers(
-  worker: Worker | undefined,
-  lifecycle: AnalyzeProjectServerLifecycle,
-  state: AnalyzeProjectServerState,
-) {
-  if (!worker) {
-    return;
-  }
-
-  worker.on('error', error => {
-    logMemoryError(error);
-    void lifecycle.shutdown('worker error');
-  });
-  worker.on('exit', code => {
-    debug(`The worker thread exited with code ${code}`);
-    if (!state.shuttingDown) {
-      void lifecycle.shutdown('worker exit');
-    }
-  });
-}
-
 export const analyzeProjectServerInternals = {
   createAnalyzeProjectStreamHandler,
   createLifecycle,
@@ -709,7 +353,7 @@ export async function startAnalyzeProjectServer(
   timeout = 0,
 ): Promise<AnalyzeProjectServerResult> {
   await logMemoryConfiguration();
-  const workerData: WorkerData = { debugMemory };
+  const workerData = { debugMemory };
   const unregisterGarbageCollectionObserver = debugMemory
     ? registerGarbageCollectionObserver()
     : () => {};
